requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.4.0
scipy>=1.7.0
scikit-image>=0.18.0
Pillow>=8.3.0
PyWavelets>=1.1.1
piexif>=1.1.3
colormath>=3.0.0
brisque>=0.0.15
ipython>=7.25.0
jupyter>=1.0.0
seaborn>=0.11.0
pandas>=1.3.0